# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Me Socorre API",
    "description": "Documentação da API da plataforma de prestação de serviços Me Socorre.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/clients/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Consulta um cliente",
        "description": "Retorna o próprio perfil do cliente. Administradores podem consultar qualquer cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do usuário cliente."
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado."
          },
          "400": {
            "description": "ID inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Usuário sem permissão para consultar o cliente."
          },
          "404": {
            "description": "Cliente não encontrado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualiza um cliente",
        "description": "Atualiza os dados do próprio cliente. Administradores podem atualizar qualquer cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do usuário cliente."
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente atualizado com sucesso."
          },
          "400": {
            "description": "ID ou dados de atualização inválidos."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Usuário sem permissão para atualizar o cliente."
          },
          "404": {
            "description": "Cliente não encontrado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2,
                    "example": "João da Silva"
                  },
                  "telefone": {
                    "type": "string",
                    "nullable": true,
                    "minLength": 8,
                    "maxLength": 30,
                    "example": "51999999999"
                  },
                  "endereco": {
                    "type": "string",
                    "nullable": true,
                    "minLength": 5,
                    "example": "Rua das Flores, 100"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Informe o token JWT retornado pelo endpoint de login."
      }
    }
  }
}